+++ docs/changelog.md
@@ -17,6 +17,7 @@ title: "Changelog"
 ## next
 
 - Remove never triggered Error `InvalidPropError`
+- Change "Options-not-for-current-type" Errors into warnings (with actual information on what options are included) [[typegoose#363](https://github.com/typegoose/typegoose/issues/363)]
 - [IC] Assing `schemaOptions` in `src/internal/schema.ts` to an black object [[typegoose#357](https://github.com/typegoose/typegoose/issues/357)]
 
 ## 7.3.4